modularize pages controller
