+++ app/views/core/pages/pages/about.html.erb
