+++ app/views/core/pages/pages/contact.html.erb
@@ -22,4 +22,4 @@
         <% end %>
     </div>
   </div>
-</div>  
+</div>
