+++ app/views/core/pages/pages/home.html.erb
