+++ config/routes.rb
@@ -7,5 +7,10 @@ Rails.application.routes.draw do
   devise_for :users
 
   root 'page#show'
-  get '/:name', to: 'page#show', constraints: { subdomain: /^[a-zA-Z0-9]*/ }
+
+  scope module: 'core' do
+    scope module: 'pages' do
+      get '/:name', to: 'pages#show', constraints: { subdomain: /^[a-zA-Z0-9]*/ }
+    end
+  end
 end
